UI Tests MBank iOS/Web

+++ src/MBank/Tests/iOS/SettingsTest.php
@@ -25,39 +25,28 @@ class SettingsTest extends \MBank\Tests\MBankiOSTestCase
      */
     public function testPublicOfferAndPrivacyPolicy()
     {
-        $wallet = $this->createWalletAndLoadDashboard();
-        $this->byName('Profile')->click();
-        $this->byName('Settings')->click();
-        // Check Public Offer Displayed
-        $this->waitForElementDisplayedByName('Public Offer');
-        $this->byName('Public Offer')->click();
-        $this->waitForElementDisplayedByName('Back to Settings icon');
-        $this->byName('Back to Settings icon')->click();
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[2]/UIAButton[2]');
-        // Check Privacy Policy Displayed
-        $this->byName('Privacy Policy')->click();
-        $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[3]/UIAWebView[1]');
-        $this->waitForElementDisplayedByName('Back to Settings icon');
-        $this->waitForElementDisplayedByName('Log out');
-        // Delete wallet
-        $this->getAPIService()->deleteWallet($wallet->phone);
-    }
-
-    /**
-     * @group Settings
-     */
-    public function testSetMail()
-    {
-        $wallet = $this->createWalletAndLoadDashboard();
-        $this->byName('Profile')->click();
-        $this->byName('Settings')->click();
-        // Set Mail
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[3]/UIATableView[1]/UIATableCell[6]/UIATextField[1]')
-             ->value('sad-nu@mail.ru');
-        $this->byXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[3]/UIAButton[1]')->click();
-        $this->byName('Menu icon')->click();
-        $this->waitForElementDisplayedByName('Your balance');
-        // Delete wallet
-        $this->getAPIService()->deleteWallet($wallet->phone);
+        if (APP_ENV == 'ios') {
+            $wallet = $this->createWalletAndLoadDashboard();
+            $this->waitForElementDisplayedByElement('Your_balance_Button');
+            $this->byElement('Profile_Button')->click();
+            $this->byElement('Settings_Button')->click();
+            // Check Public Offer Displayed
+            $this->waitForElementDisplayedByName('Public Offer');
+            $this->byName('Public Offer')->click();
+            $this->waitForElementDisplayedByName('Back to Settings icon');
+            $this->byName('Back to Settings icon')->click();
+            $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[2]/UIAButton[2]');
+            // Check Privacy Policy Displayed
+            $this->byName('Privacy Policy')->click();
+            $this->waitForElementDisplayedByXPath('//UIAApplication[1]/UIAWindow[2]/UIAScrollView[3]/UIAWebView[1]');
+            $this->waitForElementDisplayedByName('Back to Settings icon');
+            $this->waitForElementDisplayedByName('Log out');
+            // Delete wallet
+            $this->getAPIService()->deleteWallet($wallet->phone);
+        } elseif (APP_ENV == 'web')
+        {
+            //TODO for WEB_APP
+            $this->markTestSkipped("Issue not resolved for WEB_APP");
+        }
     }
 }
